rename action files
