+++ src/components/PodsList/index.jsx
@@ -2,7 +2,7 @@ import React, { Component } from 'react';
 import { bindActionCreators } from 'redux';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
-import getPodsList from '../../redux/actions/podsActions';
+import getPodsList from '../../redux/actions/pods';
 import tellAge from '../../helpers/ageUtility';
 import './PodsList.css';
 import Header from '../Header';
