+++ src/components/ProjectsList/index.jsx
@@ -6,7 +6,7 @@ import { Link } from 'react-router-dom';
 import './ProjectsList.css';
 import InformationBar from '../InformationBar';
 import Header from '../Header';
-import getUserProjects from '../../redux/actions/projectsListActions';
+import getUserProjects from '../../redux/actions/projectsList';
 import { BigSpinner } from '../SpinnerComponent';
 import ClusterCard from '../ClusterCard';
 import crane from '../../assets/images/craneLogo.png';
