+++ src/components/UserAccounts/index.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react';
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
-import getUsersList from '../../redux/actions/usersActions';
+import getUsersList from '../../redux/actions/users';
 import './UserAccounts.css';
 import Header from '../Header';
 import { BigSpinner } from '../SpinnerComponent';
