+++ src/redux/actions/deleteApp.js
